🔨 Extract question saving logic for re-use
